fix samples

+++ samples/casting/pointer/to_pointer/ptr_to_another_ptr.rs
@@ -1,7 +1,7 @@
 fn main() {
     let x = 10;
     let y = &x as *const i32;
-    let z = y as *const i64;
+    let z = y as *const u32;
     foo(z);
 }
